moved controllers to subpackage + added users controller
